10056 - What is the Probability ?
Probability has always been an integrated part of computer algorithms. Where the deterministic
algorithms have failed to solve a problem in short time, probabilistic algorithms have come to the
rescue. In this problem we are not dealing with any probabilistic algorithm. We will just try to
determine the winning probability of a certain player.
A game is played by throwing a dice like thing (it should not be assumed that it has six sides like
an ordinary dice). If a certain event occurs when a player throws the dice (such as getting a 3, getting
green side on top or whatever) he is declared the winner. There can be N such player. So the first
player will throw the dice, then the second and at last the N-th player and again the first player and
so on. When a player gets the desired event he or she is declared winner and playing stops. You will
have to determine the winning probability of one (The I-th) of these players.
Input
Input will contain an integer S (S ≤ 1000) at first, which indicates how many sets of inputs are there.
The next S lines will contain S sets of inputs. Each line contain an integer N (N ≤ 1000) which denotes
the number players, a floating point number p which indicates the probability of the happening of a
successful event in a single throw (If success means getting 3 then p is the probability of getting 3 in
a single throw. For a normal dice the probability of getting 3 is 1/6), and I (I ≤ N) the serial of the
player whose winning probability is to be determined (Serial no varies from 1 to N). You can assume
that no invalid probability (p) value will be given as input.
Output
For each set of input, output in a single line the probability of the I-th player to win. The output
floating point number will always have four digits after the decimal point as shown in the sample
output.
/*中文*/
這是一個計算機率的問題：有一群人玩擲骰子的遊戲（一般的骰子有6面，但是你不可以假設本問題的骰子一定是一般的骰子。）假如某一個人擲出一個骰子出現某一個事件（比如說得到3或者綠色面朝上......等等），我們就說這一個人贏了。

可能有N個人參與此遊戲，所以首先由第1個人先擲骰子，再來是第2個人，第3個人，.....然後輪到第N個人，接著又輪回第1個人。當有某一個人擲出某一特定事件時，他就贏了遊戲，並且遊戲結束。你的任務就是要求出某一個人（第I個人）贏得遊戲的機率。

Input

輸入的第1列有一個整數S（S<=1000），代表以下有幾組測試資料。接下來的S列每列一組測試資料，包含有3個數。第1個數為一個整數N（N<=1000），代表參加遊戲的人數。第2個數是一個浮點數p，代表擲一次骰子會出現該特定事件的機率（例如我們規定擲出3代表此特定事件，那p就是擲出3的機率，以一顆一般的骰子來說，p=1/6=0.166666）。第3個數為一個整數I（I<=N），代表要你求的是第I個人。

Output

對於每組測試資料，請輸出第I個人贏得遊戲的機率。請準確到小數點後4位。

Sample input

2
2 0.166666 1
2 0.166666 2
Sample Output

0.5455
0.4545