10170 - The Hotel with Infinite Rooms
The city of HaluaRuti has a strange hotel with infinite rooms. The groups that come to this hotel
follow the following rules:
a) At the same time only members of one group can rent the hotel.
b) Each group comes in the morning of the check-in day and leaves the hotel in the evening of the
check-out day.
c) Another group comes in the very next morning after the previous group has left the hotel.
d) A very important property of the incoming group is that it has one more member than its previous
group unless it is the starting group. You will be given the no of members of the starting group.
e) A group with n members stays for n days in the hotel. For example, if a group of four members
comes on 1st August in the morning, it will leave the hotel on 4th August in the evening and the
next group of five members will come on 5th August in the morning and stay for five days and
so on.
Given the initial group size you will have to find the group size staying in the hotel on a specified
day.
Input
The input contains round numbers S (1 ≤ S ≤ 10000) and D (1 ≤ D < 1015) in every line. S denotes
the initial size of the group and D denotes that you will have to find the group size staying in the hotel
on D-th day (starting from 1). All the input and output integers will be less than 1015. A group size S
means that on the first day a group of S members come to the hotel and stays for S days then comes
a group of S + 1 members according to the previously described rules and so on.
Output
For each line of input, print on a single line the size of the group staying in the hotel on the D-th day.
/*中文*/
有一間很奇怪的旅館，他有無限多的房間。來到這家旅館的旅行團都遵守以下的規則：

1.同一個時間只能住一個旅行團
2.每一旅行團在一早就住進旅館，然後在另一天的晚上離開。
3.上一個旅行團離開之後，隔天早上另一旅行團即住進。
4.除了第一個旅行團之外，每個剛來的旅行團的人數都比上一個旅行團多一人。
5.有n個人的旅行團會在旅館住n天。例如：有一個4個人的旅行團在8月1日上午住進旅館，他們會在8月4日晚上離開旅館。下一個有5個人的旅行團會在8月5日住進，並且停留5天。
給你第一個旅行團的人數，請你告訴我在某一天住在這家旅館的旅行團的人數有多少人。

Input

每筆測試資料一列。每列有2個整數 S（1 <= S <= 10000）和 D（1 <= D <= 1015）。S代表第一個旅行團的人數，D代表從第一個旅行團住進後的天數（從1開始算起）。

Output

對每一測試資料，請輸出從第一個旅行團住進後的第D天時，住在旅館的旅行團的人數是多少。

Sample Input

1 6
3 10
3 14
Sample Output

3
5
6