11349 - Symmetric Matrix
You‘re given a square matrix M. Elements of this matrix are Mij : {0 < i < n, 0 < j < n}. In this
problem you’ll have to find out whether the given matrix is symmetric or not.
Definition: Symmetric matrix is such a matrix that all elements of it are non-negative and symmetric
with relation to the center of this matrix. Any other matrix is considered to be non-symmetric. For
example:
M =
5 1 3
2 0 2
3 1 5
 is symmetric

M =
5 1 3
2 0 2
0 1 5
 is not symmetric, because 3 ̸= 0
All you have to do is to find whether the matrix is symmetric or not. Elements of a matrix given
in the input are −2
32 ≤ Mij ≤ 2
32 and 0 < n ≤ 100.
Input
First line of input contains number of test cases T ≤ 300. Then T test cases follow each described in
the following way. The first line of each test case contains n – the dimension of square matrix. Then
n lines follow each of then containing row i. Row contains exactly n elements separated by a space
character. j-th number in row i is the element Mij of matrix you have to process.
/*中文*/
問題
現在給你一個正方矩陣M。M這個矩陣的元素為Mij : { 0 < i < n ,0 < j < n }。在這個問題中你必須判斷這個矩陣是否對稱(symmetric)。

定義：對稱矩陣定義為內部所有的元素皆為非負且每個元素都相對於這個矩陣的中心點對稱。 其他不滿足上述條件的矩陣皆為不對稱。
你要作的只有判斷這個矩陣是否對稱。
矩陣內的元素範圍為-232 <= Mij <= 232 且 0 < n <= 100。
Input
輸入的第一列包含一個數字T<=300，代表測資的數量。之後的每一組測試資料的第一列包含一個數值n，代表這個正方矩陣的維度。接下來的n列即為這個矩陣內的元素數值，每個數字都會以一個空白字元隔開。

Output
對每一筆測試資料輸出一行"Test #t: S"，t為第幾筆測試資料的編號，而S代表"Symmetric"或"Non-symmetric"兩種字串，取決於輸入的矩陣是否為對稱矩陣。


Sample Input
2
N = 3
5 1 3
2 0 2
3 1 5
N = 3
5 1 3
2 0 2
0 1 5

Sample Output
Test #1: Symmetric.
Test #2: Non-symmetric.