10093 - An Easy Problem!
Have you heard the fact “The base of every normal number system is 10”? Of course, I am not talking
about number systems like Stern Brockot Number System. This problem has nothing to do with this
fact but may have some similarity.
You will be given an N based integer number R and you are given the guaranty that R is divisible
by (N − 1). You will have to print the smallest possible value for N. The range for N is 2 ≤ N ≤ 62
and the digit symbols for 62 based number is (0..9 and A..Z and a..z). Similarly, the digit symbols
for 61 based number system is 0..9 and A..Z and a..y) and so on.
Input
Each line in the input file will contain an integer (as defined in mathematics) number of any integer
base (2..62). You will have to determine what is the smallest possible base of that number for the given
conditions. No invalid number will be given as input.
Output
If number with such condition is not possible output the line ‘such number is impossible!’ For each
line of input there will be only a single line of output. The output will always be in decimal number
system.
給你一個N進位的整數R，並且保證R一定可以用(N-1)除的盡。你的任務是找出最小的N（2 <= N <= 62，R的長度可能到達10000位數）。在本問題中，62進位數字系統的符號為（0..9，A..Z，a..z）。類似的，61進位數字系統的符號為（0..9，A..Z，a..y）。依此類推，2進位數字系統的符號為（0..1）。

Input

每組測試資料一列，每列有一個整數R。可能包含了+、-符號。最前面也可能含有沒有意義的空白或0。請參考Sample Input。

Output

對每一組測試資料，輸出最小的N（以十進位來表示）。如果找不到，請輸出"such number is impossible!"
Sample Input
3
5
A
Sample Output
4
6
11