10019 - Funny Encryption Method
A student from ITESM Campus Monterrey plays with a new encryption method for numbers. These
method consist of the following steps:
Steps : Example
1. Read the number N to encrypt : M = 265
2. Interpret N as a decimal number : X1 = 265 (decimal)
3. Convert the decimal interpretation of N to its binary representation : X1 = 100001001 (binary)
4. Let b1 be equal to the number of 1’s in this binary representation : b1 = 3
5. Interpret N as a Hexadecimal number : X2 = 265 (hexadecimal)
6. Convert the hexadecimal interpretation of N to its binary representation : X2 = 1001100101
7. Let b2 be equal to the number of 1’s in the last binary representation : b2 = 5
8. The encryption is the result of M xor (b1 ∗ b2) : 265 xor (3*5) = 262
This student failed Computational Organization, thats why this student asked the judges of ITESM
Campus Monterrey internal ACM programming Contest to ask for the numbers of 1’s bits of this two
representations so that he can continue playing.
You have to write a program that read a Number and give as output the number b1 and b2
Input
The first line will contain a number N which is the number of cases that you have to process. Each
of the following N Lines (0 < N ≤ 1000) will contain the number M (0 < M ≤ 9999, in decimal
representation) which is the number the student wants to encrypt.
Output
You will have to output N lines, each containing the number b1 and b2 in that order, separated by one
space corresponding to that lines number to crypt
/*中文*/
一位來自墨西哥蒙特瑞技術研究學院(ITESM Campus Monterrey)的學生發表他新發明的數值加密演算法，這個方法的步驟如下：
Steps : Example

1) 假設令N = M = 265為欲加密的數字

2) 把N視為十進位數值，令X1=265(十進制)

3) 把X1由十進制轉為二進制，故X1=100001001(二進制)

4) 針對以二進制表示的X1計算共有幾個1，X1=100001001共有3個1，即令B1=3

5) 把N視為十六進位數值，令X2=265(十六進制)

6) 把X2由十六進制轉為二進制，故X2=1001100101(二進制)

7) 針對以二進制表示的X2計算共有幾個1，X2=1001100101共有5個1，即令B2=5

8) 最後的編碼為 M xor (b1*b2) M xor (3*5) = 262

這位學生在計算組識(Computational Organization)這門課被當掉了，所以他請求校方在ACM的試題上出一題計算共有幾個位元1的題目，好讓他能順利發表他的加密演算法。

Task :
你必須寫一個程式能讀入一個整數，然後輸出該整數的b1, b2值

THE INPUT
第一列為一數值N，表示共有幾組測試資料，接下來的N列(0<N<=1000)分別為一數值M(0<M<=9999, 十進制)，該數即為欲加密的數值。
THE OUTPUT
你必須輸出N列，每一列依序輸出B1與B2，以空白字元隔開。
SAMPLE INPUT
3
265
111
1234
SAMPLE OUTPUT
3 5 
6 3 
5 5