10050 - Hartals
A social research organization has determined a simple set of parameters to simulate the behavior of
the political parties of our country. One of the parameters is a positive integer h (called the hartal
parameter) that denotes the average number of days between two successive hartals (strikes) called by
the corresponding party. Though the parameter is far too simple to be flawless, it can still be used to
forecast the damages caused by hartals. The following example will give you a clear idea:
Consider three political parties. Assume h1 = 3, h2 = 4 and h3 = 8 where hi
is the hartal parameter
for party i (i = 1, 2, 3). Now, we will simulate the behavior of these three parties for N = 14 days.
One must always start the simulation on a Sunday and assume that there will be no hartals on weekly
holidays (on Fridays and Saturdays).
1 2 3 4 5 6 7 8 9 10 11 12 13 14
Days
Su Mo Tu We Th Fr Sa Su Mo Tu We Th Fr Sa
Party 1 x x x x
Party 2 x x x
Party 3 x
Hartals 1 2 3 4 5
The simulation above shows that there will be exactly 5 hartals (on days 3, 4, 8, 9 and 12) in 14
days. There will be no hartal on day 6 since it is a Friday. Hence we lose 5 working days in 2 weeks.
In this problem, given the hartal parameters for several political parties and the value of N, your
job is to determine the number of working days we lose in those N days.
Input
The first line of the input consists of a single integer T giving the number of test cases to follow.
The first line of each test case contains an integer N (7 ≤ N ≤ 3650) giving the number of days over
which the simulation must be run. The next line contains another integer P (1 ≤ P ≤ 100) representing
the number of political parties in this case. The ith of the next P lines contains a positive integer hi
(which will never be a multiple of 7) giving the hartal parameter for party i (1 ≤ i ≤ P).
Output
For each test case in the input output the number of working days we lose. Each output must be on a
separate line.
/*中文*/
社會研究組識決定採用一組簡單的參數，用來模擬我國政黨運作的行為，其中一個參數為正整數h(稱為“罷會參數(hartal parameter)”)用來表示同一個政黨連續兩次罷會的間隔時間，雖然這個參數太過簡略了，但還是可以用來預測政黨罷會所帶來的負面影響。接下來的例子會有清楚的說明：
考慮有三個政黨的情況，假設 h1 = 3, h2 = 4, h3 = 8，其中 hi 為政黨 i(i=1, 2, 3)的“罷會參數”，接下來我們摸擬這三個政黨在N=14天內的罷會行為，模擬的情況總是設定第一天為星期日，並且每週的例假日(星期五與星期六)不會有罷會的情況。

1	2	3	4	5	6	7	8	9	10	11	12	13	14
Days														
Su	Mo	Tu	We	Th	Fr	Sa	Su	Mo	Tu	We	Th	Fr	Sa
Party 1			x			x			x			x		
Party 2				x				x				x		
Party 3								x						
Hartals			1	2				3	4			5		

上述的模擬結果顯示14天內將罷會5天(第3, 4, 8, 9, 12日)，而第6天沒有罷會是因為該天是例假日星期五，因此可看出兩週的議程已去掉了5天。

本問題會給定各個黨政的“罷會參數”與總議程的天數N，你的任務是算出在N天內共有多少個工作天因為各政黨的罷會而導致議程延岩。

INPUT
輸入的第一列為一個用來表示有幾組測試資料的整數T。

每組測試資料的第一列為整數N ( $7 \le N \le 3650$) ，用來表示所模擬議程的天數。下一列為另一個整數P ( $1 \le P \le 100$)表示共有P個政黨，接下來的P列分別為各政黨的“罷會參數”(絕不會為7的倍數)。

OUTPUT
輸出的每一列表示每一組測試資料所模擬出來的罷會天數(損失多少個工作天)。

SAMPLE INPUT
2
14
3
3
4
8
100
4
12
15
25
40
SAMPLE OUTPUT
5
15