10235 - Simply Emirp
An integer greater than 1 is called a prime number if its only positive divisors (factors) are 1 and itself.
Prime numbers have been studied over the years by a lot of mathematicians. Applications of prime
numbers arise in Cryptography and Coding Theory among others.
Have you tried reversing a prime? For most primes, you get a composite (43 becomes 34). An
Emirp (Prime spelt backwards) is a Prime that gives you a different Prime when its digits are reversed.
For example, 17 is Emirp because 17 as well as 71 are Prime.
In this problem, you have to decide whether a number N is Non-prime or Prime or Emirp. Assume
that 1 < N < 1000000.
Interestingly, Emirps are not new to NTU students. We have been boarding 199 and 179 buses for
quite a long time!
Input
Input consists of several lines specifying values for N.
Output
For each N given in the input, output should contain one of the following:
1. ‘N is not prime.’, if N is not a Prime number.
2. ‘N is prime.’, if N is Prime and N is not Emirp.
3. ‘N is emirp.’, if N is Emirp.
/*中文*/
Content
一個比 1 大的整數如果只有 1 和他本身自己共 2 個因數，我們稱這個數為質數（prime number）。多年來質數一直被數學家們研究著。質數也常被應用在密碼學和編碼理論中。

那麼你曾經把質數倒轉過來嗎？對大部分的質數來說，你將會得到一個組合數（例如：43 變成 34）現在，我們要定義 Emirp（就是把 Prime 反過來拼）：如果你把一個質數反過來之後，他仍然是一個質數，並且和原來那個質數不同，那我們就稱這個數為 emirp number。例如：17 是一個emirp，因為 17 和 71 都是質數。在這個問題中，你必須要決定某一個整數 N 是非質數，質數，或 emirp。你可以假設 1<N<1000000。

Input
輸入的每一行測試資料有 1 個整數 N
Output
對每一輸入 N，輸出以下的訊息：

1. "N is not prime."，如果 N 不是一個質數

2. "N is prime."，如果 N 是一個質數，但是不是一個 Emirp

3. "N is emirp."，如果 N 是一個 emirp

Sample Input
17
18
19
179
199
131
Sample Output
17 is emirp.
18 is not prime.
19 is prime.
179 is emirp.
199 is emirp.
131 is prime.