10908 - Largest Square
Given a rectangular grid of characters you have to find out the length of a side of the largest square such
that all the characters of the square are same and the center [intersecting point of the two diagonals]
of the square is at location (r, c). The height and width of the grid is M and N respectively. Upper
left corner and lower right corner of the grid will be denoted by (0, 0) and (M − 1, N − 1) respectively.
Consider the grid of characters given below. Given the location (1, 2) the length of a side of the largest
square is 3.
abbbaaaaaa
abbbaaaaaa
abbbaaaaaa
aaaaaaaaaa
aaaaaaaaaa
aaccaaaaaa
aaccaaaaaa
Input
The input starts with a line containing a single integer T (< 21). This is followed by T test cases. The
first line of each of them will contain three integers M, N and Q (< 21) separated by a space where
M, N denotes the dimension of the grid. Next follows M lines each containing N characters. Finally,
there will be Q lines each containing two integers r and c. The value of M and N will be at most 100.
Output
For each test case in the input produce Q + 1 lines of output. In the first line print the value of M, N
and Q in that order separated by single space. In the next Q lines, output the length of a side of the
largest square in the corresponding grid for each (r, c) pair in the input.
/*中文*/
給你一個由字元構成的矩形以及座標 (r,c)，請求出由相同字元所構成的最大正方形。(r,c) 代表此最大正方形的中心，矩形的左上角座標為 (0,0)，右下角為 (M−1,N−1)。以下圖來說，給你座標 (1,2)，則此最大正方形的邊長為 3。
輸入
輸入的第一列為一整數 T (T < 21)，代表接下來要測試的組數。每組輸入的第一行包括三個整數 M、N 和 Q，M、N 指矩形的大小，接下來的 M 行包含 N 個字元，在接下來的 Q 行包含兩個整數 (r,c)。M、N 的值不超過 100。

輸出
對於每組輸出 Q+1 行。每組的第一行輸出 M、N 和 Q，接下來的 Q 行輸出以 (r,c) 為中心做出的最大正方形的邊長。

範例輸入
1
2
3
4
5
6
7
8
9
10
11
12
13
1
7 10 4
abbbaaaaaa
abbbaaaaaa
abbbaaaaaa
aaaaaaaaaa
aaaaaaaaaa
aaccaaaaaa
aaccaaaaaa
1 2
2 4
4 6
5 2
範例輸出
1
2
3
4
5
7 10 4
3
1
5
1